refactor: using jwt on create project

+++ src/main/java/kea/alog/project/common/dto/TokenPayloadDto.java
@@ -7,5 +7,5 @@ public class TokenPayloadDto {
 
     private String userNN;
     private String userEmail;
-    private int userPk;
+    private Long userPk;
 }
